+++ src/HorBorder.css
@@ -12,7 +12,3 @@
     width: 100%;
     margin-right: 10px;
 }
-
-.hb-hidden {
-    visibility: hidden;
-}
